start refactoring prepare_package.dart
