+++ dev/bots/prepare_package.dart
@@ -17,149 +17,8 @@ import 'package:platform/platform.dart' show LocalPlatform, Platform;
 import 'package:pool/pool.dart';
 import 'package:process/process.dart';
 
-const String gobMirror = 'https://flutter.googlesource.com/mirrors/flutter';
-const String githubRepo = 'https://github.com/flutter/flutter.git';
-const String mingitForWindowsUrl = 'https://storage.googleapis.com/flutter_infra_release/mingit/'
-    '603511c649b00bbef0a6122a827ac419b656bc19/mingit.zip';
-const String releaseFolder = '/releases';
-const String gsBase = 'gs://flutter_infra_release';
-const String gsReleaseFolder = '$gsBase$releaseFolder';
-const String baseUrl = 'https://storage.googleapis.com/flutter_infra_release';
-const int shortCacheSeconds = 60;
-const String frameworkVersionTag = 'frameworkVersionFromGit';
-const String dartVersionTag = 'dartSdkVersion';
-const String dartTargetArchTag = 'dartTargetArch';
-
-/// Exception class for when a process fails to run, so we can catch
-/// it and provide something more readable than a stack trace.
-class PreparePackageException implements Exception {
-  PreparePackageException(this.message, [this.result]);
-
-  final String message;
-  final ProcessResult? result;
-  int get exitCode => result?.exitCode ?? -1;
-
-  @override
-  String toString() {
-    String output = runtimeType.toString();
-    output += ': $message';
-    final String stderr = result?.stderr as String? ?? '';
-    if (stderr.isNotEmpty) {
-      output += ':\n$stderr';
-    }
-    return output;
-  }
-}
-
-enum Branch {
-  beta,
-  stable,
-  master,
-  main;
-}
-
-/// A helper class for classes that want to run a process, optionally have the
-/// stderr and stdout reported as the process runs, and capture the stdout
-/// properly without dropping any.
-class ProcessRunner {
-  ProcessRunner({
-    ProcessManager? processManager,
-    this.subprocessOutput = true,
-    this.defaultWorkingDirectory,
-    this.platform = const LocalPlatform(),
-  }) : processManager = processManager ?? const LocalProcessManager() {
-    environment = Map<String, String>.from(platform.environment);
-  }
-
-  /// The platform to use for a starting environment.
-  final Platform platform;
-
-  /// Set [subprocessOutput] to show output as processes run. Stdout from the
-  /// process will be printed to stdout, and stderr printed to stderr.
-  final bool subprocessOutput;
-
-  /// Set the [processManager] in order to inject a test instance to perform
-  /// testing.
-  final ProcessManager processManager;
-
-  /// Sets the default directory used when `workingDirectory` is not specified
-  /// to [runProcess].
-  final Directory? defaultWorkingDirectory;
-
-  /// The environment to run processes with.
-  late Map<String, String> environment;
-
-  /// Run the command and arguments in `commandLine` as a sub-process from
-  /// `workingDirectory` if set, or the [defaultWorkingDirectory] if not. Uses
-  /// [Directory.current] if [defaultWorkingDirectory] is not set.
-  ///
-  /// Set `failOk` if [runProcess] should not throw an exception when the
-  /// command completes with a non-zero exit code.
-  Future<String> runProcess(
-    List<String> commandLine, {
-    Directory? workingDirectory,
-    bool failOk = false,
-  }) async {
-    workingDirectory ??= defaultWorkingDirectory ?? Directory.current;
-    if (subprocessOutput) {
-      stderr.write('Running "${commandLine.join(' ')}" in ${workingDirectory.path}.\n');
-    }
-    final List<int> output = <int>[];
-    final Completer<void> stdoutComplete = Completer<void>();
-    final Completer<void> stderrComplete = Completer<void>();
-    late Process process;
-    Future<int> allComplete() async {
-      await stderrComplete.future;
-      await stdoutComplete.future;
-      return process.exitCode;
-    }
-
-    try {
-      process = await processManager.start(
-        commandLine,
-        workingDirectory: workingDirectory.absolute.path,
-        environment: environment,
-      );
-      process.stdout.listen(
-        (List<int> event) {
-          output.addAll(event);
-          if (subprocessOutput) {
-            stdout.add(event);
-          }
-        },
-        onDone: () async => stdoutComplete.complete(),
-      );
-      if (subprocessOutput) {
-        process.stderr.listen(
-          (List<int> event) {
-            stderr.add(event);
-          },
-          onDone: () async => stderrComplete.complete(),
-        );
-      } else {
-        stderrComplete.complete();
-      }
-    } on ProcessException catch (e) {
-      final String message = 'Running "${commandLine.join(' ')}" in ${workingDirectory.path} '
-          'failed with:\n$e';
-      throw PreparePackageException(message);
-    } on ArgumentError catch (e) {
-      final String message = 'Running "${commandLine.join(' ')}" in ${workingDirectory.path} '
-          'failed with:\n$e';
-      throw PreparePackageException(message);
-    }
-
-    final int exitCode = await allComplete();
-    if (exitCode != 0 && !failOk) {
-      final String message = 'Running "${commandLine.join(' ')}" in ${workingDirectory.path} failed';
-      throw PreparePackageException(
-        message,
-        ProcessResult(0, exitCode, null, 'returned $exitCode'),
-      );
-    }
-    return utf8.decoder.convert(output).trim();
-  }
-}
+import 'prepare_package/common.dart';
+import 'prepare_package/process_runner.dart';
 
 typedef HttpReader = Future<Uint8List> Function(Uri url, {Map<String, String> headers});
